Initialized bottom appbar add setting section add pengurusan section Initialized web version refactoring some code
