+++ lib/home/page/tetapan_view.dart
@@ -1,0 +1,59 @@
+import 'package:flutter/material.dart';
+
+class TetapanView extends StatelessWidget {
+  const TetapanView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: CustomScrollView(
+        slivers: [
+          SliverAppBar.large(
+            title: const Text('Tetapan'),
+          ),
+          SliverList(
+            delegate: SliverChildListDelegate(
+              [
+                Padding(
+                  padding: const EdgeInsets.symmetric(horizontal: 10.0),
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(
+                        'Paparan',
+                        style: TextStyle(
+                          color: Theme.of(context).colorScheme.primary,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                      const SizedBox(height: 10),
+                      ListTile(
+                        leading: Icon(Icons.format_paint),
+                        title: Text('Tema'),
+                        subtitle: Text('Cerah'),
+                        onTap: () {},
+                      ),
+                      Text(
+                        'Akaun',
+                        style: TextStyle(
+                          color: Theme.of(context).colorScheme.primary,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                      ListTile(
+                        leading: Icon(Icons.format_paint),
+                        title: Text('Log Masuk'),
+                        subtitle: Text('Log masuk akaun pengurusan'),
+                        onTap: () {},
+                      ),
+                    ],
+                  ),
+                ),
+              ],
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
